Clean DccWalletInfo
